+++ src/Core/Managed/Shared/Extensibility/SdkVersionPropertyTelemetryInitializer.cs
@@ -3,12 +3,12 @@
     using System.Linq;
     using System.Reflection;
     using System.Threading;
-    using Microsoft.ApplicationInsights.DataContracts;
+    using Channel;
 
     /// <summary>
     /// Initializes SDK Properties: SDK Version and SDKMode.
     /// </summary>
-    internal sealed class SdkVersionPropertyContextInitializer : IContextInitializer
+    internal sealed class SdkVersionPropertyTelemetryInitializer : ITelemetryInitializer
     {
         private const string SDKVersion = "SDKVersion";
         private string sdkVersion;
@@ -16,24 +16,24 @@
         /// <summary>
         /// Adds a telemetry property for the version of SDK.
         /// </summary>
-        public void Initialize(TelemetryContext context)
+        public void Initialize(ITelemetry telemetry)
         {
             var version = LazyInitializer.EnsureInitialized(ref this.sdkVersion, this.GetAssemblyVersion);
-            if (string.IsNullOrEmpty(context.Internal.SdkVersion))
+            if (string.IsNullOrEmpty(telemetry.Context.Internal.SdkVersion))
             {
-                context.Internal.SdkVersion = version;
+                telemetry.Context.Internal.SdkVersion = version;
             }
         }
 
         private string GetAssemblyVersion()
         {
 #if !WINRT && !CORE_PCL && !UWP
-            return typeof(SdkVersionPropertyContextInitializer).Assembly.GetCustomAttributes(false)
+            return typeof(SdkVersionPropertyTelemetryInitializer).Assembly.GetCustomAttributes(false)
                     .OfType<AssemblyFileVersionAttribute>()
                     .First()
                     .Version;
 #else
-            return typeof(SdkVersionPropertyContextInitializer).GetTypeInfo().Assembly.GetCustomAttributes<AssemblyFileVersionAttribute>()
+            return typeof(SdkVersionPropertyTelemetryInitializer).GetTypeInfo().Assembly.GetCustomAttributes<AssemblyFileVersionAttribute>()
                     .First()
                     .Version;
 #endif